Update README.md
Create a classification model in Python. 

Analyse the model and provide insight into which features/variables influence the outcome of the classification the most: on a global level, and specifically for observation #4 and #20.

Develop some form of plot or graph with brief comments on  observations.

Evaluate the relevance of potential features, both on a global and a local level. [SHAP & LIME]